+++ scripts/lichessTools-init.js
@@ -7,6 +7,7 @@ lichessTools.loadTool(LiChessTools.Tools.FixWakeLock);
 lichessTools.loadTool(LiChessTools.Tools.FixAbortController);
 lichessTools.loadTool(LiChessTools.Tools.EmitRedraw);
 lichessTools.loadTool(LiChessTools.Tools.EmitChapterChange);
+lichessTools.loadTool(LiChessTools.Tools.EmitPuzzleChange);
 lichessTools.loadTool(LiChessTools.Tools.TranspositionBehavior);
 lichessTools.loadTool(LiChessTools.Tools.RandomVariation);
 lichessTools.loadTool(LiChessTools.Tools.DetectThirdParties);