+++ tools/shapeRank/tool.js
@@ -34,13 +34,13 @@
     ensureShapeRank=()=>{
       const parent=this.lichessTools;
       const analysis=parent.lichess.analysis;
-      const shapeRankEnabled=parent.currentOptions.getValue('shapeRank');
       this.chessground=analysis?.chessground || $('div.cg-wrap.lichessTools-boardOverlay')[0]?.chessground;
       const drawable=this.chessground?.state.drawable;
       if (!drawable) return;
 
       const reshape=($this,result,shapes)=>{
         this.clearRankShapes(shapes);
+        if (!this.options.enabled) return;
         const dict={}
         const drawnShapes = [];
         let rank=0;
@@ -50,6 +50,7 @@
           const rankShape={
             type: 'rank',
             orig: shape.orig,
+            dest: false, // fix lichess bug where this is found as the shape to erase
             customSvg:parent.makeSvg('<text x="10%" y="50%" font-size="200%" fill="black" stroke="'+shape.brush+'">'+rank+'</text>',this.chessground)
           };
           dict[rankShape.orig]=true;
@@ -61,7 +62,7 @@
         }
       };
 
-      if (shapeRankEnabled) {
+      if (this.options.enabled) {
         if (!parent.isWrappedFunction(drawable.onChange,'shapeRank')) {
           drawable.onChange=parent.wrapFunction(drawable.onChange,{
             id:'shapeRank',
@@ -75,18 +76,18 @@
       } 
     };
 
-    waitForChessground=(value)=>{
+    waitForChessground=()=>{
       const parent=this.lichessTools;
       const lichess=parent.lichess;
       const analysis=lichess?.analysis;
       this.chessground=analysis?.chessground || $('div.cg-wrap.lichessTools-boardOverlay')[0]?.chessground;
       if (!this.chessground) {
-        parent.global.setTimeout(this.waitForChessground.bind(this),1000);
+        parent.global.setTimeout(this.waitForChessground,500);
         return;
       }
       lichess.pubsub.off('shapeRank',this.ensureShapeRank);
       lichess.pubsub.off('redraw',this.ensureShapeRank);
-      if (value) {
+      if (this.options.enabled) {
         lichess.pubsub.on('shapeRank',this.ensureShapeRank);
         lichess.pubsub.on('redraw',this.ensureShapeRank);
         parent.global.setTimeout(this.ensureShapeRank,500); //TODO without the timeout something clears the shapes in about 250ms at first page load (probably a web socket event)
@@ -99,6 +100,7 @@
     async start() {
       const parent=this.lichessTools;
       const value=parent.currentOptions.getValue('shapeRank');
+      this.options={ enabled: value };
       this.logOption('Show the order of arrows and circles', value);
       await this.waitForChessground(value);
     }